+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/request/UserLoginDTO.java
@@ -1,0 +1,18 @@
+package com.ivanovp.medical_record.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserLoginDTO {
+
+    @NotBlank(message = "Username is required")
+    private String username;
+
+    @NotBlank(message = "Password is required")
+    private String password;
+}